refactor(playerAnalysis): split the summary's body, routes, and tiebreaker
AnalysisSummary.tsx was 375 lines holding four unrelated jobs. Every
declaration moves verbatim so `git log -M` follows it, and the file keeps
its default export alone at 54 lines.

analysisParts.tsx takes the pieces with no analysis logic in them.
mondayNight.tsx takes the tiebreaker builders. AnalysisRoutes.tsx takes
`Routes` under a name that does not read as routing in a react-router
app. AnalysisBody.tsx takes the switch on `result.kind`.

Each new file rendering an `analysis__*` class imports the stylesheet
itself, following the precedent at Standing.tsx:8.

No measurable change. AnalysisSummary.test.tsx is unchanged and its 43
cases pass.
