+++ src/components/playerAnalysis/CLAUDE.md
@@ -8,7 +8,12 @@ their name in either table.
   `PlayerStatusIcon`. `useArrival` takes a name handed in from a table.
 - `PlayerAnalysisDialog.scss`: the status hues alone. Everything else about the look
   comes from `components/dialog/`.
-- `AnalysisSummary`: renders `PlayerAnalysis`: the pick grid, the routes, the
-  `MNF Points` section. Decides once whether the week is done, for both halves.
+- `AnalysisSummary`: the standing above the body. Decides once whether the week
+  is done, for both halves.
+- `analysisParts`: `Section`, `Picks`, `Message`, the pieces with no analysis
+  logic of their own, shared by the three files below.
+- `mondayNight`: the MNF Points tiebreaker, its sentence and its route line.
+- `AnalysisRoutes`: the paths list, folded until asked to show more.
+- `AnalysisBody`: the switch on `PlayerAnalysis.kind` that picks what to render.
 - `Standing`: where the picked player stands, read off the scores, so it shows
   while their routes are worked out.
